src/node/node_ref.rs: added `NodeRef::replace_with_html`; src/dom_tree.rs: fixed `Tree::append_children_from_another_tree`
